调整approve代码结构
