+++ mapper/mainmapper.go
@@ -2,13 +2,11 @@ package mapper
 
 import (
 	"github.com/QOSGroup/qbase/mapper"
-	"github.com/QOSGroup/qbase/store"
 	"github.com/tendermint/tendermint/crypto"
 )
 
 const (
-	BaseMapperName = "basemapper"
-	storeKey       = "base"
+	storeKey = "base"
 )
 
 type MainMapper struct {
@@ -17,11 +15,11 @@ type MainMapper struct {
 
 func NewMainMapper() *MainMapper {
 	var baseMapper = MainMapper{}
-	baseMapper.BaseMapper = mapper.NewBaseMapper(store.NewKVStoreKey(storeKey))
+	baseMapper.BaseMapper = mapper.NewBaseMapper(nil, storeKey)
 	return &baseMapper
 }
 
-func GetBaseMapperStoreKey() string {
+func GetMainStoreKey() string {
 	return storeKey
 }
 
@@ -29,10 +27,6 @@ func (mapper *MainMapper) Copy() mapper.IMapper {
 	cpyMapper := &MainMapper{}
 	cpyMapper.BaseMapper = mapper.BaseMapper.Copy()
 	return cpyMapper
-}
-
-func (mapper *MainMapper) Name() string {
-	return BaseMapperName
 }
 
 // 保存CA
