+++ txs/approve/mapper.go
@@ -1,17 +1,14 @@
-package mapper
+package approve
 
 import (
 	"fmt"
 	"github.com/QOSGroup/qbase/mapper"
-	"github.com/QOSGroup/qbase/store"
 	btypes "github.com/QOSGroup/qbase/types"
-	"github.com/QOSGroup/qos/types"
 )
 
 const (
-	ApproveMapperName = "approvemapper"
-	approveStoreKey   = "approve"
-	approveKey        = "from:[%s]/to:[%s]"
+	approveStoreKey = "approve"
+	approveKey      = "from:[%s]/to:[%s]"
 )
 
 type ApproveMapper struct {
@@ -20,7 +17,7 @@ type ApproveMapper struct {
 
 func NewApproveMapper() *ApproveMapper {
 	var approveMapper = ApproveMapper{}
-	approveMapper.BaseMapper = mapper.NewBaseMapper(store.NewKVStoreKey(approveStoreKey))
+	approveMapper.BaseMapper = mapper.NewBaseMapper(nil, approveStoreKey)
 	return &approveMapper
 }
 
@@ -39,28 +36,24 @@ func (mapper *ApproveMapper) Copy() mapper.IMapper {
 	return approveMapper
 }
 
-func (mapper *ApproveMapper) Name() string {
-	return ApproveMapperName
-}
-
 // 获取授权
-func (mapper *ApproveMapper) GetApprove(from btypes.Address, to btypes.Address) (types.Approve, bool) {
-	approve := types.NewApprove(from, to, nil, nil)
+func (mapper *ApproveMapper) GetApprove(from btypes.Address, to btypes.Address) (Approve, bool) {
+	approve := NewApprove(from, to, nil, nil)
 	key := BuildApproveKey(from.String(), to.String())
 	exists := mapper.BaseMapper.Get(key, &approve)
 	return approve, exists
 }
 
 // 保存授权
-func (mapper *ApproveMapper) SaveApprove(approve types.Approve) error {
+func (mapper *ApproveMapper) SaveApprove(approve Approve) error {
 	key := BuildApproveKey(approve.From.String(), approve.To.String())
 	mapper.BaseMapper.Set(key, approve)
 	return nil
 }
 
 // 删除授权
-func (mapper *ApproveMapper) DeleteApprove(approve types.ApproveCancel) error {
-	key := BuildApproveKey(approve.From.String(), approve.To.String())
+func (mapper *ApproveMapper) DeleteApprove(from btypes.Address, to btypes.Address) error {
+	key := BuildApproveKey(from.String(), to.String())
 	mapper.BaseMapper.Del(key)
 	return nil
 }
