+++ txs/codec.go
@@ -3,6 +3,7 @@ package txs
 import (
 	"github.com/QOSGroup/qbase/baseabci"
 	"github.com/QOSGroup/qos/account"
+	"github.com/QOSGroup/qos/txs/approve"
 	"github.com/tendermint/go-amino"
 )
 
@@ -14,12 +15,8 @@ func init() {
 }
 
 func RegisterCodec(cdc *amino.Codec) {
+	approve.RegisterCodec(cdc)
 	cdc.RegisterConcrete(&TxCreateQSC{}, "qos/txs/TxCreateQSC", nil)
 	cdc.RegisterConcrete(&TxIssueQsc{}, "qos/txs/TxIssueQsc", nil)
 	cdc.RegisterConcrete(&TransferTx{}, "qos/txs/TransferTx", nil)
-	cdc.RegisterConcrete(&ApproveCreateTx{}, "qos/txs/ApproveCreateTx", nil)
-	cdc.RegisterConcrete(&ApproveIncreaseTx{}, "qos/txs/ApproveIncreaseTx", nil)
-	cdc.RegisterConcrete(&ApproveDecreaseTx{}, "qos/txs/ApproveDecreaseTx", nil)
-	cdc.RegisterConcrete(&ApproveUseTx{}, "qos/txs/ApproveUseTx", nil)
-	cdc.RegisterConcrete(&ApproveCancelTx{}, "qos/txs/ApproveCancelTx", nil)
 }
